Add capability for "toggle" type buttons.
A new attribute "actions_type" has been added to the widgets.
When set to "state", the action in the "actions" array
corresponding to the current state will be run.

You can now add the "content" attribute to any widget's "state_style"
entries which will actually change the content of the widget
depending on the state.

A good example of using both capabilities is an ATX Power toggle
button where clicking it turns the power on or off depnding on
it's current state. The default config file has a new button for
just that.

If you have multiple actions in a widget and want a dialog to pop
up asking which to run, set "actions_type" to "choose".

If you have multiple actions and set "actions_type" to "all" or
don't specify it at all, all actions will run in sequence.

Also...

The settings toggle buttons have been updated to use the new
"actions_type" and this changed the values stored in the settings
cookie.  You'll probably be prompted to turn polling on again
the next time you run DueUI.

The browser title is now automatically set to "DueUI" and the
hostname of your Duet.