Fix handling of "macro" action
The macro action was double-prepending macro names with "/macros"

Now the "macro" action can take EITHER a "file" or "macro"
parameter.  If "file", you must supply the full path to the
macro file such as "/macros/mymacro".  If "macro", you must
supply a file name relative to the "/macros" directory such as
"mymacro" which will automatically be converted to "/macros/mymacro".